Removed OpenVR from project.

+++ src/Mesh.cpp
@@ -154,7 +154,7 @@ void dg::Mesh::AddTriangle(Vertex v1, Vertex v2, Vertex v3, Winding winding) {
       if (!!(attributes & Flag::TANGENT)) {
         vertexTangents.push_back(v[i]->data.tangent);
       }
-      index = vertexPositions.size() - 1;
+      index = (unsigned int)vertexPositions.size() - 1;
       vertexMap[hash] = index;
     } else {
       index = pair->second;
@@ -763,7 +763,7 @@ void dg::DirectXMesh::Draw() const {
   Graphics::Instance->context->IASetIndexBuffer(
     indexBuffer, DXGI_FORMAT_R32_UINT, 0);
 
-  Graphics::Instance->context->DrawIndexed(indices.size(), 0, 0);
+  Graphics::Instance->context->DrawIndexed((UINT)indices.size(), 0, 0);
 }
 
 #pragma endregion
